/kbin RTR#15 Preparing for the refactoring, UserManager (#1265)
Reviewed-on: https://codeberg.org/Kbin/kbin-core/pulls/1265
Co-committed-by: Ernest Wiśniewski <ernestwisniewski2@gmail.com>

+++ src/Kbin/Magazine/ModeratorRequest/MagazineModeratorRequestToggle.php
@@ -4,17 +4,15 @@ declare(strict_types=1);
 
 namespace App\Kbin\Magazine\ModeratorRequest;
 
-use App\DTO\ModeratorDto;
 use App\Entity\Magazine;
+use App\Entity\ModeratorRequest;
 use App\Entity\User;
-use App\Kbin\Magazine\Moderator\MagazineAddModerator;
 use App\Repository\ModeratorRequestRepository;
 use Doctrine\ORM\EntityManagerInterface;
 
-readonly class MagazineAcceptModeratorRequest
+readonly class MagazineModeratorRequestToggle
 {
     public function __construct(
-        private MagazineAddModerator $magazineAddModerator,
         private ModeratorRequestRepository $moderatorRequestRepository,
         private EntityManagerInterface $entityManager,
     ) {
@@ -22,14 +20,21 @@ readonly class MagazineAcceptModeratorRequest
 
     public function __invoke(Magazine $magazine, User $user): void
     {
-        ($this->magazineAddModerator)(new ModeratorDto($magazine, $user));
-
         $request = $this->moderatorRequestRepository->findOneBy([
             'magazine' => $magazine,
             'user' => $user,
         ]);
 
-        $this->entityManager->remove($request);
+        if ($request) {
+            $this->entityManager->remove($request);
+            $this->entityManager->flush();
+
+            return;
+        }
+
+        $request = new ModeratorRequest($magazine, $user);
+
+        $this->entityManager->persist($request);
         $this->entityManager->flush();
     }
 }
